requirements.txt
PyGithub>=2.3.0
setuptools>=68.2.0